Update 04-special-considerations.Rmd
Thank you for including information about transparency. TN struggles in that arena.

Look at all sentences in the published format that begin with an "A." It messes with the spacing and looks weird. Might want to consider adding an extra space in the coding based on this font oddity. W and Y also have a less extreme version of the same problem.

There seem to be multiple authors across different sections and some writing norms (educational data scientists vs. data scientists in education) are established in one chapter then changed in the next. Consider scanning the text for commonly used phrases, labels, etc. and making them consistent.